% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_by.R
\name{min_by}
\alias{min_by}
\title{Compute minimum by group}
\usage{
min_by(df, var, by)
}
\arguments{
\item{df}{A data frame}

\item{var}{Variable(s) to compute minimum of.}

\item{by}{Variables to group by.}
}
\value{
A data frame.
}
\description{
Quickly and easily compute the minimum value by group
}
\examples{
mtcars |> min_by(mpg, cyl)
}
