% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_mpg_by_cyl}
\alias{min_mpg_by_cyl}
\title{Minimum mpg broken down by cylinder}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\usage{
min_mpg_by_cyl
}
\description{
\url{https://twitter.com/zx8754/status/1597551262077419526}
}
\examples{
min_mpg_by_cyl
}
\keyword{datasets}
